ATRIAN/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
pytest>=8.0.0
tqdm>=4.66.0